requirements.txt
PyPDF2
Faker
PyYAML
openai
pdfrw
